requirements.txt
python-dotenv
openai
requests  # Required for Veo-3 and RunwayML backends
google-auth
google-auth-oauthlib